Rename MST to Prims file
Prims is an algorithm to find a Minimum Spanning tree given a graph.
In this case, the graph is provided in the form of an adjacency matrix,
and my algorithm returns a list of tuples that represent the edges of
the MST (V1, V2, weight).
This Implementation uses the heapq library to make the algorithm run in
O(Nlog(n)) time.